backend/requirements.txt
# Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2

# Database
SQLAlchemy==2.0.21

# Environment Variables
python-dotenv==1.0.0

# Security
Werkzeug==2.3.7

# ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Image Processing (if needed for SVG/DICOM)
Pillow==10.0.0
opencv-python==4.8.0.76

# Medical Image Processing (optional)
pydicom==2.4.3

# API and Validation
marshmallow==3.20.1

# Development Tools
python-dateutil==2.8.2
